loom/requirements.txt
daft[all]==0.4.9
kuzu==0.7.1
chromadb==0.5.20
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pymupdf==1.24.10
python-docx==1.1.2
spacy==3.7.6
pyarrow==17.0.0
pandas==2.2.3
nbformat==5.10.4
tree-sitter==0.23.0
pyyaml==6.0.2
python-multipart==0.0.12
slowapi==0.1.9
python-jose==3.3.0
rclone-python==0.2.3
pillow==10.4.0
tqdm==4.66.5
watchdog==4.0.2
schedule==1.2.2
